get-ghe-access-token/requirements.txt
cryptography==37.0.2
PyJWT==2.4.0
requests==2.28.1
